+++ internal/testdata/common.go
@@ -25,40 +25,40 @@ const (
 
 func initResourceAttributes1(dest pcommon.Map) {
 	dest.Clear()
-	dest.InsertString("resource-attr", "resource-attr-val-1")
+	dest.UpsertString("resource-attr", "resource-attr-val-1")
 }
 
 func initSpanEventAttributes(dest pcommon.Map) {
 	dest.Clear()
-	dest.InsertString("span-event-attr", "span-event-attr-val")
+	dest.UpsertString("span-event-attr", "span-event-attr-val")
 }
 
 func initSpanLinkAttributes(dest pcommon.Map) {
 	dest.Clear()
-	dest.InsertString("span-link-attr", "span-link-attr-val")
+	dest.UpsertString("span-link-attr", "span-link-attr-val")
 }
 
 func initMetricExemplarAttributes(dest pcommon.Map) {
 	dest.Clear()
-	dest.InsertString("exemplar-attachment", "exemplar-attachment-value")
+	dest.UpsertString("exemplar-attachment", "exemplar-attachment-value")
 }
 
 func initMetricAttributes1(dest pcommon.Map) {
 	dest.Clear()
-	dest.InsertString("label-1", "label-value-1")
+	dest.UpsertString("label-1", "label-value-1")
 }
 
 func initMetricAttributes12(dest pcommon.Map) {
 	initMetricAttributes1(dest)
-	dest.InsertString(testLabelKey2, testLabelValue2)
+	dest.UpsertString(testLabelKey2, testLabelValue2)
 }
 
 func initMetricAttributes13(dest pcommon.Map) {
 	initMetricAttributes1(dest)
-	dest.InsertString("label-3", "label-value-3")
+	dest.UpsertString("label-3", "label-value-3")
 }
 
 func initMetricAttributes2(dest pcommon.Map) {
 	dest.Clear()
-	dest.InsertString(testLabelKey2, testLabelValue2)
+	dest.UpsertString(testLabelKey2, testLabelValue2)
 }