+++ internal/testdata/log.go
@@ -50,8 +50,8 @@ func fillLogOne(log plog.LogRecord) {
 	log.SetTraceID([16]byte{0x08, 0x04, 0x02, 0x01})
 
 	attrs := log.Attributes()
-	attrs.InsertString("app", "server")
-	attrs.InsertInt("instance_num", 1)
+	attrs.UpsertString("app", "server")
+	attrs.UpsertInt("instance_num", 1)
 
 	log.Body().SetStringVal("This is a log message")
 }
@@ -63,8 +63,8 @@ func fillLogTwo(log plog.LogRecord) {
 	log.SetSeverityText("Info")
 
 	attrs := log.Attributes()
-	attrs.InsertString("customer", "acme")
-	attrs.InsertString("env", "dev")
+	attrs.UpsertString("customer", "acme")
+	attrs.UpsertString("env", "dev")
 
 	log.Body().SetStringVal("something happened")
 }